Update README.md
Change node and nginx to always restart. There is some magic sauce that makes the nginx container stop when the unit goes away. Could not see what. Instead added '-f' to force removal of any existing containers on start. Added an ExecStop to stop the service when the unit is stopped.